requirements.txt
gradio
torch
torchvision
transformers
gtts
pillow




